Changed inverter loop detection logic
Previously, a single net from the inverter chain was used to detect loops.
However, if that net was outside of the loop, the program would recursively call
itself up to net_count times. This could lead to OOM errors.

This change keeps a set of nets (the first and each 1024th net thereafter) to
check for loops.

Signed-off-by: Mitch Bailey <d.mitch.bailey@gmail.com>